restored app.py & routes.py
