+++ app/routes.py
@@ -1,0 +1,23 @@
+from flask import Flask, render_template
+
+app = Flask(__name__)
+
+
+@app.route("/")
+def index():
+    return render_template("index.html")
+
+
+@app.route("/login")
+def login():
+    return render_template("login.html")
+
+
+@app.route("/how-to-play")
+def how_to_play():
+    return render_template("howtoplay.html")
+
+
+@app.route("/leaderboard")
+def leaderboard():
+    return render_template("leaderboard.html")
